+++ obj/src/services/HttpEndpoint.js
@@ -61,6 +61,7 @@ class HttpEndpoint {
         this._counters = new pip_services3_components_node_2.CompositeCounters();
         this._maintenanceEnabled = false;
         this._fileMaxSize = 200 * 1024 * 1024;
+        this._protocolUpgradeEnabled = false;
         this._registrations = [];
     }
     /**
@@ -86,6 +87,7 @@ class HttpEndpoint {
         this._connectionResolver.configure(config);
         this._maintenanceEnabled = config.getAsBooleanWithDefault('options.maintenance_enabled', this._maintenanceEnabled);
         this._fileMaxSize = config.getAsLongWithDefault('options.file_max_size', this._fileMaxSize);
+        this._protocolUpgradeEnabled = config.getAsBooleanWithDefault('options.protocol_upgrade_enabled', this._protocolUpgradeEnabled);
     }
     /**
      * Sets references to this endpoint's logger, counters, and connection resolver.
@@ -154,10 +156,11 @@ class HttpEndpoint {
                     options.certificate = certificate;
                     //options.ca = ca;
                 }
-                // Create instance of express application   
+                options.handleUpgrades = this._protocolUpgradeEnabled;
+                // Create instance of restify application   
                 let restify = require('restify');
                 this._server = restify.createServer(options);
-                // Configure express application
+                // Configure restify application
                 this._server.use(restify.plugins.acceptParser(this._server.acceptable));
                 //this._server.use(restify.authorizationParser());
                 //this._server.use(restify.CORS());